Revert "Localisation updates from http://translatewiki.net."
That commit removed message 'visualeditor-help-tool' for some reason.
Could be a race condition (relatively new message), but unlikely as other
translations of that message were not removed and are still in the file.

Anyway, reverting so they can try again. And to avoid this messages's
translations from being deleted next time translatewiki reads this from git.

This reverts commit 73f829904688fa729001976eda8f85213f37ace2.

Change-Id: If457b61e3fd3703a59ee67a9a0c26b0c9a36d0d4