Rename 'formulæ' to 'formulae'.
Everyone loves a ligature, but within our codebase we should
be consistent, and sadly that means using en-US, not en-GB-archaic.

Also having non-basic latin characters in filenames is a bit
of a pain for developers.

Similarly for the i18n message, the ligature should go in
MessagesEn_gb if anywhere.

Also in this commit, re-saved the BF thumbnails as plain SVG.

Change-Id: I21a2bf040bed132d3f1c73c916c731d36e0ad852