Put the shim from 9abad038 back (was reverted in b9bbf60f)
The shim triggers bug 58249, but that's not as bad as we thought.
It turns out LocalisationUpdate still finishes despite the scary
exception, it just skips VisualEditor and finishes the rest successfully.

I will work on making the LU error more useful, and eventually
making it support JSON blobs, but the shim isn't causing a
lot of damage for now.

Change-Id: I92f11524d891e901f1a089dd8cad1232478a5ee4