Only allow pointer events on shields inside generated content nodes
Often the wrapper and contents have different size and positioning
characteristics than the shields, and shouldn't be interacted with
directly. For example, a centered image has a 100% width wrapper with a
figure inside with a fixed width and horizontal margins set to auto. The
wrapper ends up being a click block to the left and right, when all we
really want to do is make the shields clickable.

Change-Id: I2c2c592d78025cfd983c20d263a0e600616c9ae1