jsduck: Have a default --meta-tags setting
Our custom bin/generateDocs will override accordingly, but
for production we want to default to jsduck 4. That way
we can use the generic macro instead of a custom job
to run the repo's bin/generateDocs.

We need that for mwext-VisualEditor because that also
needs other overrides, but for VisualEditor standalone
we can do plain jsduck.

Change-Id: I09507b048f1f487e3067853de959a56ab4d08723