Add placeholders to MediaSizeWidget
Add the ability to adjust input placeholders in media size widget.
This is particularly aimed at the new 'default size' fix in ve-mw.
Unfortunately, we usually don't have the placeholder details when
the input widgets are instantiated, so the method added makes sure
these are cached, appear on 0 values and can be retrieved.

Also, the valid dimension check method was adapted to accomodate
a situation where the widget is empty but its placeholders have
data.

Change-Id: I27d5efda12b7e57291211269f1f5fbc610aef4ad