Update OOjs UI to v0.1.0-pre (68c91ae065)
Includes the great window refactor of 2014.

Also adjust ve.ui.Inspector and ve.ui.Dialog for the signature
changes in their parent constructors, but keep the windowSet
parameter for now. We use it to extract the surface and we should
probably do this differently, but for now this will keep things
working.

It turned out that passing in a windowSet in the place of the
config object actually worked surprisingly well, because it has
a .$ property and the icon and title fall back to static properties.
However, the media insertion dialog (footless) and the save dialog
(small) were broken.

Change-Id: I52e6a28e5a16e8806b9c155392e8775f68afa03a